workspace/python/agents/pack-rag/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "使用 Vertex AI RAG 引擎回答與 Vertex AI 文件相關的問題"
authors = [{ name = "Lastingyeh", email = "lastingyeh@gmail.com" }]
license = { text = "Apache License 2.0" }
readme = "README.md"
requires-python = ">=3.11,<3.13"

# --- 核心依賴 ---
dependencies = [
    "google-cloud-aiplatform[adk,agent-engines]>=1.108.0", # Google Cloud AI 平台
    "google-adk>=1.10.0",                                  # Google ADK 框架
    "pydantic-settings>=2.8.1",                            # 設定管理
    "tabulate>=0.9.0",                                     # 表格格式化
    "google-auth>=2.36.0",                                 # 身份認證
    "requests>=2.32.3",                                    # HTTP 請求
    "llama-index>=0.12",                                   # LlamIndex 引擎
    "openinference-instrumentation-google-adk>=0.1.0",     # 可觀測性工具
    "openinference-instrumentation>=0.1.34",
    "arize-otel>=0.8.2",                                   # 監控與追蹤
    "arize>=7.36.0",                                       # Arize 平台整合
]

# --- 可選依賴（程式碼品質） ---
[project.optional-dependencies]
lint = [
    "ruff>=0.4.6",                     # 快速 Python linter
    "mypy>=1.15.0",                    # 靜態型別檢查
    "codespell>=2.2.0",                # 拼字檢查
    "types-pyyaml>=6.0.12.20240917",
    "types-requests>=2.32.0.20240914",
]

# --- 開發依賴 ---
[dependency-groups]
dev = [
    "pytest>=8.3.5",              # 測試框架
    "pytest-mock>=3.14.0",        # Mock 工具
    "scikit-learn>=1.6.1",        # 機器學習庫
    "pytest-cov>=6.0.0",          # 測試覆蓋率
    "pytest-asyncio>=0.26.0",     # 非同步測試支援
    "agent-starter-pack>=0.14.1", # Agent 啟動模板
]

# --- CI/CD 與測試中繼資料 ---
[tool.agent-starter-pack.testing]
RAG_CORPUS = "projects/adk-devops/locations/us-central1/ragCorpora/456"
ENV = "test"

# --- Ruff 程式碼檢查設定 ---
[tool.ruff]

[tool.ruff.lint]
extend-select = ["I"]     # 啟用 isort 相容性檢查
ignore = ["E501", "C901"] # 忽略行長度與複雜度警告

[tool.ruff.lint.isort]
known-first-party = ["rag"] # 定義本地套件

# --- MyPy 型別檢查設定（嚴格模式） ---
[tool.mypy]
disallow_untyped_calls = true                                     # 禁止呼叫無型別註解的函式
disallow_untyped_defs = true                                      # 禁止未標註型別的定義
disallow_incomplete_defs = true                                   # 禁止不完整的型別定義
no_implicit_optional = true                                       # 禁止隱含的 Optional
check_untyped_defs = true                                         # 檢查未標註的定義
disallow_subclassing_any = true                                   # 禁止繼承 Any
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
follow_imports = "silent"
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc", "no-untyped-call", "no-any-return"]
exclude = [".venv"]

# --- 拼字檢查設定 ---
[tool.codespell]
ignore-words-list = "rouge"
skip = "./locust_env/*,uv.lock,.venv,./frontend,**/*.ipynb"

# --- Pytest 設定 ---
[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"

# --- 建置系統 ---
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# --- Agent 啟動包配置 ---
[tool.agent-starter-pack]
example_question = "Alphabet 2024 年 10-K 報告中提到的主要業務部門有哪些？"

[tool.agent-starter-pack.settings]
agent_directory = "rag"
